break out front-end validation
